pyproject.toml
[project]
name = "ginga-api"
version = "0.1.0"
description = "Ginga REST API (FastAPI, Cognito, PostgreSQL async)"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "python-dotenv>=1.0.0",
    "sqlalchemy[asyncio]>=2.0.36",
    "asyncpg>=0.30.0",
    "alembic>=1.14.0",
    "greenlet>=3.1.0",
    "pydantic>=2.10.0",
    "pydantic-settings>=2.6.0",
    "requests>=2.32.0",
    "python-multipart>=0.0.12",
    "python-jose[cryptography]>=3.3.0",
    "boto3>=1.35.0",
    "mangum>=0.19.0",
]

[dependency-groups]
dev = [
    "httpx>=0.28.0",
    "pytest>=8.3.0",
    "pytest-asyncio>=0.24.0",
    "ruff>=0.8.0",
]

[tool.ruff]
line-length = 100
target-version = "py312"
src = ["app"]

[tool.ruff.lint]
select = ["E", "F", "I", "UP"]

[tool.ruff.lint.per-file-ignores]
"app/database/migrations/env.py" = ["E402"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "function"
